String ctors, unittest++, scripts++
- oz
  * arrays++:
    + value type in aSet, aIndex ... is arbitrary
    + aFind(), aFindLast()
  * misc++ in iterables & arrays
  * Log::putsRaw()
  * String ctors for char, byte, ubyte ...
- unittest
  * separate unittest subfolder
  * all tests updated & improved
- scripts
  * updated and cleaned up, some comments added
  * building now takes place in build/<PLATFORM>
  * support-all.sh copies libraries into support/<PLATFORM>
  * launcher scripts -> etc/
- CMakeFiles.txt fixed for CPack

+++ src/oz/Alloc.cc
@@ -88,13 +88,13 @@ void Alloc::printSummary()
   Log::indent();
 
   Log::println( "current chunks     %d", count );
-  Log::println( "current amount     %.2f MiB (%ld B)",
+  Log::println( "current amount     %.2f MiB (%lu B)",
                 float( amount ) / ( 1024.0f * 1024.0f ), ulong( amount ) );
   Log::println( "maximum chunks     %d", maxCount );
-  Log::println( "maximum amount     %.2f MiB (%ld B)",
+  Log::println( "maximum amount     %.2f MiB (%lu B)",
                 float( maxAmount ) / ( 1024.0f * 1024.0f ), ulong( maxAmount ) );
   Log::println( "cumulative chunks  %d", sumCount );
-  Log::println( "cumulative amount  %.2f MiB (%ld B)",
+  Log::println( "cumulative amount  %.2f MiB (%lu B)",
                 float( sumAmount ) / ( 1024.0f * 1024.0f ), ulong( sumAmount ) );
 
   Log::unindent();
@@ -114,7 +114,7 @@ void Alloc::printLeaks()
 
   bt = firstObjectTraceEntry;
   while( bt != null ) {
-    Log::println( "Leaked object at %p of size %d B allocated", bt->address, int( bt->size ) );
+    Log::println( "Leaked object at %p of size %lu B allocated", bt->address, ulong( bt->size ) );
     Log::indent();
     Log::printTrace( bt->stackTrace );
     Log::unindent();
@@ -124,7 +124,7 @@ void Alloc::printLeaks()
 
   bt = firstArrayTraceEntry;
   while( bt != null ) {
-    Log::println( "Leaked array at %p of size %d B allocated", bt->address, int( bt->size ) );
+    Log::println( "Leaked array at %p of size %lu B allocated", bt->address, ulong( bt->size ) );
     Log::indent();
     Log::printTrace( bt->stackTrace );
     Log::unindent();
